refactor: rename classes folder
