+++ src/app.ts
@@ -1,3 +1,3 @@
-import { Bot } from './classes/bot';
+import { Bot } from './structures/bot';
 
 void new Bot().start();
