+++ src/commands/music/resume.ts
@@ -1,5 +1,5 @@
 import { Queue } from 'discord-music-player';
-import { Command, CommandModule } from '../../classes/command';
+import { Command, CommandModule } from '../../structures/command';
 import { generateEmbed, replyEmbed, replyMessage } from '../../utils';
 import { musicValidator } from '../../utils/validators';
