+++ src/events/guild/guildDelete.ts
@@ -1,5 +1,5 @@
 import { Guild } from 'discord.js';
-import { Bot } from '../../classes/bot';
+import { Bot } from '../../structures/bot';
 import { RunEvent } from '../../interfaces';
 import { prisma } from '../../services';
