+++ src/events/user/voiceStateUpdate.ts
@@ -1,6 +1,6 @@
 import { Guild } from '@prisma/client';
 import { VoiceState } from 'discord.js';
-import { Bot } from '../../classes/bot.js';
+import { Bot } from '../../structures/bot';
 import { RunEvent } from '../../interfaces';
 import { prisma } from '../../services';
 import { generateEmbed, sendEmbedToChannelArr } from '../../utils';
