+++ src/structures/bot.ts
