+++ src/structures/stream.ts
